Create tests for header level detection (failing) [skip ci]
