+++ tests/prepare_release/changelog/header_level_detection/CHANGELOG.md
@@ -1,0 +1,7 @@
+# CHANGELOG
+
+# 1.0.0
+
+## Breaking Changes 🍳
+
+### Well for one, we're using a second first-level header 🙀
