+++ tests/prepare_release/changelog/header_level_detection/Cargo.toml
@@ -1,0 +1,2 @@
+[package]
+version = "1.0.0"
