+++ tests/prepare_release/changelog/header_level_detection/EXPECTED_CHANGELOG.md
@@ -1,0 +1,13 @@
+# CHANGELOG
+
+# 1.1.0 ([DATE])
+
+## Features
+
+### We support custom header levels now 🎉
+
+# 1.0.0
+
+## Breaking Changes 🍳
+
+### Well for one, we're using a second first-level header 🙀
