+++ tests/prepare_release/changelog/header_level_detection/dry_run_output.txt
@@ -1,0 +1,16 @@
+Would add the following to Cargo.toml: 1.1.0
+Would add the following to CHANGELOG.md:
+# 1.1.0 ([DATE])
+
+## Features
+
+### We support custom header levels now 🎉
+
+Would add files to git:
+  Cargo.toml
+  CHANGELOG.md
+
+Would create a release on GitHub with name 1.1.0 ([DATE]) and tag v1.1.0 and body:
+## Features
+
+### We support custom header levels now 🎉
